Create vulnFTPServer.c
This is the main file for the fake FTP server which presents as an FTP server but which is vulnerable to buffer overflows in various pages. In order to have it work, it's also required to compile a DLL file without any security so it always loads in the same spot.